+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 **Implemented enhancements:**
 
-- Handle ValueError with an error code and error message in RAMSTKSimilarItem.set\_attributes\(\). [\#252](https://github.com/ReliaQualAssociates/ramstk/issues/252)
+-  Handle ValueError with an error code and error message in RAMSTKSimilarItem.set\_attributes\(\). [\#252](https://github.com/ReliaQualAssociates/ramstk/issues/252)
 - Move Similar Item Work View \_do\_request\_edit\_function\(\) method to a stand-alone assistant [\#212](https://github.com/ReliaQualAssociates/ramstk/issues/212) [[module:similaritem](https://github.com/ReliaQualAssociates/ramstk/labels/module:similaritem)]
 - Refactor \_on\_combo\_changed\(\) and \_on\_focus\_out\(\) methods in component workviews to eliminate duplicate code. [\#208](https://github.com/ReliaQualAssociates/ramstk/issues/208) [[module:hardware](https://github.com/ReliaQualAssociates/ramstk/labels/module:hardware)]
 - Refactor workviews to have \_\_make\_ui\(\), \_\_set\_properties\(\), \_\_set\_callbacks\(\) methods [\#207](https://github.com/ReliaQualAssociates/ramstk/issues/207) [[module:fmea](https://github.com/ReliaQualAssociates/ramstk/labels/module:fmea)] [[module:pof](https://github.com/ReliaQualAssociates/ramstk/labels/module:pof)]
@@ -19,6 +19,7 @@
 - PyGTK deprecation [\#58](https://github.com/ReliaQualAssociates/ramstk/issues/58)
 - Support Python 3 [\#57](https://github.com/ReliaQualAssociates/ramstk/issues/57)
 - Add target to docs/Makefile to auto publish gh-pages [\#223](https://github.com/ReliaQualAssociates/ramstk/pull/223) ([weibullguy](https://github.com/weibullguy))
+- Refactor MIL-HDBK-217F part stress methods [\#220](https://github.com/ReliaQualAssociates/ramstk/pull/220) [[module:hardware](https://github.com/ReliaQualAssociates/ramstk/labels/module:hardware)] ([weibullguy](https://github.com/weibullguy))
 - Move docs to GitHub Pages [\#217](https://github.com/ReliaQualAssociates/ramstk/pull/217) ([weibullguy](https://github.com/weibullguy))
 - Refactor piQ assignment for parts count analyses [\#216](https://github.com/ReliaQualAssociates/ramstk/pull/216) [[module:hardware](https://github.com/ReliaQualAssociates/ramstk/labels/module:hardware)] ([weibullguy](https://github.com/weibullguy))
 - Issue \#204: Remove duplicate code in overstressed methods [\#214](https://github.com/ReliaQualAssociates/ramstk/pull/214) [[module:hardware](https://github.com/ReliaQualAssociates/ramstk/labels/module:hardware)] ([weibullguy](https://github.com/weibullguy))